Move chart data into a separate export for use elsewhere
